chore(desktop): split SettingsPage into section components
Extract 5 memoized section components from SettingsPage (741→103 lines)
and deduplicate format constants shared across 3 files.
